requirements.txt
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
arch==8.0.0
attrs==26.1.0
Authlib==1.6.9
babel==2.18.0
beartype==0.22.9
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
cyclopts==4.10.1
dash==4.1.0
dash-bootstrap-components==2.0.4
dateparser==1.4.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
Farama-Notifications==0.0.4
fastapi==0.135.3
fastmcp==3.2.0
filelock==3.25.2
flake8==7.3.0
Flask==3.1.3
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gevent==25.9.1
greenlet==3.3.2
gymnasium==1.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jeepney==0.9.0
Jinja2==3.1.6
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
kiwisolver==1.5.0
lightgbm==4.6.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
mcp==1.27.0
mcp-server==0.1.4
mdurl==0.1.2
more-itertools==11.0.1
mpmath==1.3.0
msgpack==1.0.2
multidict==6.7.1
narwhals==2.18.1
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
openapi-pydantic==0.5.1
opentelemetry-api==1.40.0
packaging==26.0
pandas==3.0.2
pathable==0.5.0
patsy==1.0.2
pika==1.3.2
pillow==12.2.0
platformdirs==4.9.4
playwright==1.58.0
plotly==6.6.0
pluggy==1.6.0
ppi_client==1.2.4
prometheus_client==0.25.0
propcache==0.4.1
psycopg2-binary==2.9.11
py-key-value-aio==0.4.4
pycodestyle==2.14.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyee==13.0.1
pyflakes==3.4.0
Pygments==2.20.0
PyJWT==2.12.1
PyMySQL==1.1.2
pyparsing==3.3.2
pyperclip==1.11.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python_binance==1.0.36
pytz==2026.1.post1
PyYAML==6.0.3
QuantLib==1.41
redis==7.4.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
retrying==1.4.2
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
scipy==1.17.1
seaborn==0.13.2
SecretStorage==3.5.0
setuptools==70.2.0
signalrcorePPI==0.9.2
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
sse-starlette==3.3.4
stable_baselines3==2.8.0
starlette==1.0.0
statsmodels==0.14.6
sympy==1.14.0
ta==0.11.0
tabulate==0.10.0
torch==2.11.0+cpu
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uncalled-for==0.2.0
urllib3==2.6.3
uvicorn==0.43.0
vcrpy==8.1.1
watchfiles==1.1.1
websocket-client==1.0.0
websockets==16.0
Werkzeug==3.1.8
wrapt==2.1.2
yarl==1.23.0
zipp==3.23.0
zope.event==6.1
zope.interface==8.2
